Fixed include big merdas

+++ Classes/Location.cpp
@@ -10,7 +10,7 @@ void Location::setLocationAddress(Address address) {
     this->address = address;
 }
 
-void Location::setLocationCoords(pair<int, int> coords) {
+void Location::setLocationCoords(pair<float, float> coords) {
     this->coords = coords;
 }
 
@@ -19,7 +19,7 @@ Address Location::getLocationAddress() const {
     return address;
 }
 
-pair<int, int> Location::getLocationCoords() const {
+pair<float, float> Location::getLocationCoords() const {
     return coords;
 }
